requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
tensorflow==2.13.0
ccxt==3.1.54
matplotlib==3.7.2
ta==0.10.2
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==1.0.0
joblib==1.3.1
websockets==11.0.3
yarl==1.9.2
pymongo==4.4.1
pyyaml==6.0.1
pytest==7.4.0
loguru==0.7.0
MetaTrader5==5.0.45
talib-binary==0.4.24
backtrader==1.9.76.123
aiohttp==3.8.5
requests==2.31.0
plotly==5.15.0
Flask==2.3.2
jinja2==3.1.2
jwt==1.3.1
# Dependencies for network discovery and one-tap trading
miniupnpc==2.2.5
qrcode==7.4.2
pyzbar==0.1.9
pillow==10.0.0
netifaces==0.11.0
psutil==5.9.5
zeroconf==0.58.0
